housekeeping: refactor login screen into one file
